Update 'toMatrix' implementations to match feature's type.

+++ src-js/naive_bayes/bernoulli/estimator_providers_map/bernoulli_f64_u64_provider.ts
@@ -22,7 +22,7 @@ class BernoulliNBF64U64Provider
   }
 
   toMatrix(x: DenseMatrix): DenseMatrixRs {
-    return x.asRsMatrix('f32')
+    return x.asRsMatrix('f64')
   }
 
   deserialize(data: Buffer): BernoulliNBF64U64 {
